pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-prophet"
version = "0.1.10"
description = "A minimal LLM forecasting agent scaffolding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "trafilatura",
    "pydantic>=2.0",
    "typer",
    "rich",
    "tenacity",
    "pyyaml",
    "platformdirs",
    "litellm",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "perplexityai",
    "exa-py",
    "tavily-python",
]
perplexity = [
    "perplexityai",
]
exa = [
    "exa-py",
]
tavily = [
    "tavily-python",
]
hf = [
    "datasets",
]

[project.scripts]
miniprophet = "miniprophet.run.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_api: tests that perform real external API calls (disabled unless explicitly selected).",
]

[tool.coverage.run]
source = ["miniprophet"]
omit = [
    "src/miniprophet/__main__.py",
    "src/miniprophet/cli/*",
    "src/miniprophet/run/app.py",
    "src/miniprophet/run/cli.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
